Moved Reader code to an abstract class . because i hate code duplication
