+++ Genie/Models/Concrete/View.cs
@@ -3,8 +3,9 @@ using Genie.Models.Abstract;
 
 namespace Genie.Models.Concrete
 {
-    internal class View: IView
+    internal class View : IView
     {
+        public string Comment { get; set; }
         public string FieldName { get; set; }
         public string Name { get; set; }
         public List<ISimpleAttribute> Attributes { get; set; }
